+++ example/index.tsx
@@ -20,6 +20,8 @@ import {
   TableDemo,
   DragResizeHorizontalDemo,
   DragResizeVerticalDemo,
+  TooltipDemo,
+  RangeDemo,
 } from './components';
 import './global-styles.scss';
 import { Login } from './pages/Login';
@@ -39,6 +41,7 @@ const App = () => (
         <Route path="/AutocompleteDemo" component={AutocompleteDemo} />
         <Route path="/ButtonDemo" component={ButtonDemo} />
         <Route path="/progress" component={ProgressDemo} />
+        <Route path="/range" component={RangeDemo} />
         <Route path="/toggle" component={ToggleDemo} />
         <Route path="/select" component={FormSelectDemo} />
         <Route path="/multiUpload" component={MultiUploadDemo} />
@@ -47,11 +50,9 @@ const App = () => (
         <Route path="/tabGroup" component={TabGroupDemo} />
         <Route path="/copyToClipboard" component={CopyToClipboardDemo} />
         <Route path="/table" component={TableDemo} />
-        <Route
-          path="/dragResizeHorizontal"
-          component={DragResizeHorizontalDemo}
-        />
+        <Route path="/dragResizeHorizontal" component={DragResizeHorizontalDemo}/>
         <Route path="/dragResizeVertical" component={DragResizeVerticalDemo} />
+        <Route path="/tooltipDemo" component={TooltipDemo} />
       </Switch>
     </BrowserRouter>
   </div>