+++ src/formSelect/FormSelect.tsx
@@ -53,6 +53,15 @@ export type FormSelectProps = {
    * select style
    */
   style?: any;
+  /**
+   * you can add an option that will have the specified label but an empty value
+   * nullOption will be selected by default
+   */
+  nullOption?: string;
+  /**
+   * select value which is programatically added by user
+   */
+  value?: number | string;
 };
 
 export const FormSelect = ({
@@ -61,6 +70,7 @@ export const FormSelect = ({
   optionGroups,
   options = [],
   onChange,
+  value,
   id,
   ariaLabel,
   label,
@@ -68,6 +78,7 @@ export const FormSelect = ({
   hint,
   error,
   style,
+  nullOption,
 }: FormSelectProps) => {
   const sharedOptions: OptionProps[] | undefined = optionGroups
     ? optionGroups.flatMap((group: OptionGroupProps) => [...group.options])
@@ -81,14 +92,17 @@ export const FormSelect = ({
     option => option.selected
   );
 
-  const initialValue: string =
-    preselectedValue !== undefined
-      ? preselectedValue.value
-      : options.length
-      ? options[0].value
-      : '';
+  const initialValue: string | number = value
+    ? value
+    : nullOption !== undefined
+    ? nullOption
+    : preselectedValue !== undefined
+    ? preselectedValue.value
+    : options.length
+    ? options[0].value
+    : '';
 
-  const [value, setValue] = useState<string>(initialValue);
+  const [selectValue, setSelectValue] = useState<string | number>(initialValue);
 
   const getOptions = (options: OptionProps[]): JSX.Element[] =>
     options.map((item: OptionProps, index: number) => (
@@ -101,7 +115,7 @@ export const FormSelect = ({
     ));
 
   const handleChange = (event: ChangeEvent<HTMLSelectElement>) => {
-    setValue(event.target.value);
+    setSelectValue(event.currentTarget.value);
     if (onChange) onChange(event);
   };
 
@@ -115,7 +129,7 @@ export const FormSelect = ({
       {hint && <Hint {...hint} />}
       {error && <ErrorMessage {...error} />}
       <select
-        value={value}
+        value={selectValue}
         name={name || 'formSelect'}
         id={id || 'formSelect'}
         className={className}
@@ -123,6 +137,7 @@ export const FormSelect = ({
         onChange={handleChange}
         style={style}
       >
+        {nullOption && <Option value="" label={nullOption} selected={true} />}
         {getOptions(options)}
         {optionGroups && getOptionGroups(optionGroups)}
       </select>