+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [Accessibility issues fixed against all current components](https://github.com/Capgemini/dcx-react-library/issues/200)
 - [toolTip](https://github.com/Capgemini/dcx-react-library/issues/199)
 - [table with custom rows](https://github.com/Capgemini/dcx-react-library/issues/221)
+- [range](https://github.com/Capgemini/dcx-react-library/issues/198)
 
 ## 0.2.4 (06/10/21)